[Refactor/PyTorch] Refactor pytorch subplugin to remove .c/.cc wrapper
This commit refactors pytorch tensor-filter subplugin to remove
.c/.cc wrapper.
