+++ ext/nnstreamer/tensor_filter/meson.build
@@ -100,10 +100,7 @@ if get_option('enable-tensorflow-lite')
 endif
 
 if get_option('enable-pytorch')
-  filter_sub_torch_sources = [
-    'tensor_filter_pytorch.c',
-    'tensor_filter_pytorch_core.cc'
-  ]
+  filter_sub_torch_sources = ['tensor_filter_pytorch.cc']
 
   nnstreamer_filter_torch_sources = []
   foreach s : filter_sub_torch_sources
